Add cancel migration functionality
This commit adds the functionality to cancel a
storage live migration. It will revert the VM
volumes back to the original source volumes.

If the migration is successfully cancelled, the
migration resources will be cleaned up. If the
migration fails due to the cancellation, the
migration resources will NOT be cleaned up.

Signed-off-by: Alexander Wels <awels@redhat.com>